requirements.txt
click==7.1.2
numpy==1.19.2
pandas==1.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0click==7.1.2
IMDbPY @ git+https://github.com/alberanid/imdbpy@bfdbdd085a551a4f355ff7557a364e27e8f892af
lxml==4.5.2
numpy==1.19.2
pandas==1.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
click==7.1.2
IMDbPY @ git+https://github.com/alberanid/imdbpy@bfdbdd085a551a4f355ff7557a364e27e8f892af
lxml==4.5.2
numpy==1.19.2
pandas==1.1.2
prompt-toolkit==1.0.14
Pygments==2.7.1
PyInquirer==1.0.3
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
six==1.15.0
SQLAlchemy==1.3.19
wcwidth==0.2.5
